+++ src/game/hitbox.cpp
@@ -12,23 +12,23 @@ bool Hitbox::collides(Hitbox& hitbox) {
 }
 
 std::tuple<util::X, util::Y> Hitbox::getMaximumMovement(util::X stepX, util::Y stepY, Hitbox otherHitbox) {
-    double retX = stepX, retY = stepY;
+    int retX = stepX, retY = stepY;
 
     // If X direction was already colliding last step then we are parallel in this direction
     // I.e. do no change speed
     if (!collidesX(otherHitbox)) {
-        if (stepX > 0.0) {
+        if (stepX > 0) {
             retX -= right_ + stepX - otherHitbox.left_;
-        } else if (stepX < 0.0) {
+        } else if (stepX < 0) {
             retX += otherHitbox.right_ - (left_ + stepX);
         }
     }
 
     // Same for Y
     if (!collidesY(otherHitbox)) {
-        if (stepY > 0.0) {
+        if (stepY > 0) {
             retY -= bottom_ + stepY - otherHitbox.top_;
-        } else if (stepY < 0.0) {
+        } else if (stepY < 0) {
             retY += otherHitbox.bottom_ - (top_ + stepY);
         }
     }
